Unit tests [stage-20]
